requirements.txt
appdirs==1.4.3
click==6.7
decorator==4.0.11
Flask==0.12.2
Flask-HTTPAuth==3.2.2
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pbr==3.0.1
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.10
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.2
